ETR-Few-Shot-Evaluation/pyproject.toml
[project]
name = "etr-fr-few-shot-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "cookiecutter>=2.6.0",
    "datasets>=3.3.2",
    "einops>=0.8.1",
    "evaluate>=0.4.3",
    "faiss-cpu>=1.10.0",
    "flake8>=7.2.0",
    "jsonargparse>=4.38.0",
    "jupyter>=1.1.1",
    "lab>=8.4",
    "mypy>=1.15.0",
    "nltk>=3.9.1",
    "numpy>=2.2.3",
    "pytest>=8.3.5",
    "ray==2.32.0",
    "scipy>=1.15.2",
    "sentence-transformers>=4.1.0",
    "textacy>=0.13.0",
    "textdescriptives",
    "torch>=2.6.0",
    "transformers>=4.51.2",
    "wandb>=0.19.9",
]

[tool.uv.sources]
textdescriptives = { git = "https://github.com/FrLdy/TextDescriptives.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src/icl",
  "src/etr_fr"
]
